Abstracted the google drive picker
